Open Shell Buffer in Selected Window
The problem I've had with the shell command is that I tend to select the window
I want the shell buffer to open up, but it opens up in a new window, or a
different open window. This change opens the shell buffer in the window I select
it to open in.