Have Eshell Run Most Commands
Some commands check the terminal type before paging stuff. However,
eshell is not bound to a terminal, so it gives a weird message to the
shell, and the tries to page the output using less.

This change sets the PAGER variable to cat, which allows the command
that use a pager to show the full text to Eshell.