Customize 'display-buffer-alist' and Toggle Side Windows
Customizing the 'display-buffer-alist' variable allows me to to select
which frame or window to pull up a buffer. For example, shell is
opened in the same window I was on, whereas opening a C++ file opens
on whichever frame doesn't have a buffer with shell mode.

This is meant to be an effort in optimizing my workflow