Separate Source Blocks from Headings with Empty Line
In the last change I edited the face of the code blocks to have a
background color.  Not having a line between the blocks and the
headings causes the collapsed view to include the face properties of
the source block, aka the background color.  This change gets rid of
it.