Add Keybinding to Create TODO in tasks file
Now that I'm experimenting with encrypted files, I see that the
functionality isn't the best when wanting to capture ideas - I have to
retype my password after every entry using org capture.

This diff adds logic to create a task within the org file, with a
level 1 heading.  It also adds a timestamp for when that task was
created.