Add More to Org Agenda
Now that I'm getting more invested in the productivity system I'm
trying out,  I'm trying to pretty it up.

This diff adds specific sections for the goals and the projects I'm
working on.  The idea is to start scheduling when projects are due,
and to always have in mind how many projects I'm taking on, and what
my goals are.